support to internalAG

+++ 6-cleanUp/GCP/deleteServices.sh
@@ -8,6 +8,25 @@ export PROJECTID=$(cat ../../2-cloudServices/GCP/servicesInfo.json | jq -r '.pro
 
 gcloud container clusters delete $K8SCLUSTERNAME --zone $ZONEK8SCL --quiet
 
-gcloud compute instances delete $VMNAME --zone $ZONEVM --quiet
+if [[ $VMNAME != "" ]]; then
+    gcloud compute instances delete $VMNAME --zone $ZONEVM --quiet
+fi
 
 gcloud projects delete $PROJECTID --quiet
+
+INFO=$(
+    cat <<EOF
+    {
+        "zonek8sCl": "$ZONEK8SCL",
+        "zoneVM": "$ZONEVM",
+        "k8sClusterName": "",
+        "vmName": "",
+        "projectID": ""
+    }
+EOF
+)
+
+FILE=../../2-cloudServices/GCP/servicesInfo.json
+rm $FILE 2>/dev/null
+
+echo $INFO | jq -r '.' >>$FILE
